backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
chromadb
ultralytics
sentence-transformers
Pillow
numpy
google-generativeai